% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuttlefish}
\alias{cuttlefish}
\title{Cuttlefish Sightings Dataset}
\format{
A data frame with 750 rows and 9 variables
}
\source{
https://www.ala.org.au
}
\usage{
cuttlefish
}
\description{
Sightings of cuttlefish downloaded from https://www.ala.org.au
}
\details{
\itemize{
\item id: Record ID given by Atlas
\item name: Scientific name
\item lon, lat: Spatial coordinates of the sighting
\item date, year, month, wday, time: Date and time of sighting
\item source: Original provider of the record
\item stnid: Station ID for the nearest weather station
}
}
\examples{
data(cuttlefish)
head(cuttlefish)
}
\keyword{datasets}
