% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_koala_data.R
\name{get_koala_data}
\alias{get_koala_data}
\title{Load Koala Data}
\usage{
get_koala_data()
}
\value{
A data frame containing koala occurrence records.
}
\description{
This function loads the koala occurrence dataset from the package's `inst/extdata/` folder.
}
\examples{
\dontrun{
koala_data <- get_koala_data()
}
}
