% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koala_data.R
\docType{data}
\name{koala_data}
\alias{koala_data}
\title{Koala Occurrence Data (2023-2024)}
\format{
A data frame with the following columns:
\describe{
  \item{eventDate}{Date of observation (Date object)}
  \item{decimalLatitude}{Latitude in decimal degrees (numeric, double)}
  \item{decimalLongitude}{Longitude in decimal degrees (numeric, double)}
  \item{individualCount}{Number of individuals observed (numeric, double; may contain NA values)}
  \item{stateProvince}{Australian state where observation occurred (character)}
  \item{habitat_type}{Type of habitat (character, from ALA's cl1048 field)}
  \item{conservation_status}{Conservation status in the state (character)}
  \item{conservation_source}{Legislative source for conservation status (character)}
}
}
\source{
\url{https://www.ala.org.au}
}
\usage{
koala_data
}
\description{
This dataset contains koala occurrence records from 2023-2024,
filtered from the Atlas of Living Australia.
}
\keyword{datasets}
