% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_lga}
\alias{oz_lga}
\title{Local Government Areas (LGA) Polygons for Australia}
\format{
A spatial polygons object
}
\usage{
oz_lga
}
\description{
A spatial polygons object representing the boundaries of local government areas (LGAs) in Australia.
}
\examples{
data(oz_lga)
head(oz_lga)
}
\keyword{datasets}
