% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Nearest Weather Stations for Cuttlefish Records}
\format{
A data frame with 750 rows and 3 variables
}
\usage{
stations
}
\description{
A dataset of the nearest weather stations corresponding to each record in the cuttlefish sightings data.
}
\details{
\itemize{
\item stnid: Station ID
\item lon, lat: Spatial coordinates of the station
}
}
\examples{
data(stations)
head(stations)
}
\keyword{datasets}
