% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tourism counts}
\alias{Tourism counts}
\alias{tourism}
\title{tourism}
\format{
A 1228000 x 4 numeric array
}
\usage{
tourism
}
\description{
Quarterly records of tourist counts in regions across the country
}
\details{
\itemize{
\item Quarter
\item Region
\item Purpose
\item Trips
}
}
\examples{

library(tsibble)
data(tourism)
tourism <- tourism |>
  mutate(Quarter = yearquarter(myd(Quarter))) |>
  as_tsibble(key = c(Region, Purpose), index = Quarter))

}
\keyword{datasets}
