% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourism_activity}
\alias{tourism_activity}
\title{Monthly Tourism Trips by Stopover Activity}
\format{
A tibble with 320,188 rows and 6 variables:
* **year**: Calendar year, from 2014 to 2024.
* **month**: Ordered factor giving the month returned from the trip.
* **sa2_code_tra**: Integer SA2 code used by Tourism Research Australia.
  This can be joined to \code{\link{tourism_region}}.
* **stop_activity**: Integer activity code that can be joined to
  \code{\link{tourism_activity_name}}.
* **trips**: Estimated number of overnight trips, in thousands.
* **n_sample**: Number of survey responses supporting the estimate.
}
\usage{
tourism_activity
}
\description{
Estimated domestic overnight trips to Australian SA2 areas by month and
stopover activity. The data contain only observations with non-zero trip
estimates and non-zero survey sample sizes.
}
\details{
Records for which either \code{trips} or \code{n_sample} equals zero are
excluded. Estimates with small sample sizes should be interpreted with
appropriate caution.
}
\examples{
data(tourism_activity)
head(tourism_activity)
}
\references{
Tourism Research Australia: \url{https://www.tra.gov.au}
}
\seealso{
\code{\link{tourism_activity_name}},
  \code{\link{tourism_reason}}, \code{\link{tourism_region}}
}
\keyword{datasets}
