% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourism_activity_name}
\alias{tourism_activity_name}
\title{Tourism Stopover Activity Names}
\format{
A tibble with 8 rows and 2 variables:
* **stop_activity**: Integer stopover activity code.
* **name**: Descriptive name of the stopover activity.
}
\usage{
tourism_activity_name
}
\description{
A lookup table matching the stopover activity codes in
\code{\link{tourism_activity}} to descriptive names.
}
\examples{
data(tourism_activity_name)
tourism_activity_name
}
\references{
Tourism Research Australia: \url{https://www.tra.gov.au}
}
\seealso{
\code{\link{tourism_activity}}
}
\keyword{datasets}
