% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourism_reason}
\alias{tourism_reason}
\title{Monthly Tourism Trips by Stopover Reason}
\format{
A tibble with 196,044 rows and 6 variables:
* **year**: Calendar year, from 2014 to 2024.
* **month**: Ordered factor giving the month returned from the trip.
* **sa2_code_tra**: Integer SA2 code used by Tourism Research Australia.
  This can be joined to \code{\link{tourism_region}}.
* **stop_reason**: Integer reason code that can be joined to
  \code{\link{tourism_reason_name}}.
* **trips**: Estimated number of overnight trips, in thousands.
* **n_sample**: Number of survey responses supporting the estimate.
}
\usage{
tourism_reason
}
\description{
Estimated domestic overnight trips to Australian SA2 areas by month and
stopover reason. The data contain only observations with non-zero trip
estimates and non-zero survey sample sizes.
}
\details{
Records for which either \code{trips} or \code{n_sample} equals zero are
excluded. Estimates with small sample sizes should be interpreted with
appropriate caution.
}
\examples{
data(tourism_reason)
head(tourism_reason)
}
\references{
Tourism Research Australia: \url{https://www.tra.gov.au}
}
\seealso{
\code{\link{tourism_reason_name}},
  \code{\link{tourism_activity}}, \code{\link{tourism_region}}
}
\keyword{datasets}
