% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourism_reason_name}
\alias{tourism_reason_name}
\title{Tourism Stopover Reason Names}
\format{
A tibble with 7 rows and 2 variables:
* **stop_reason**: Integer stopover reason code.
* **name**: Descriptive name of the stopover reason.
}
\usage{
tourism_reason_name
}
\description{
A lookup table matching the stopover reason codes in
\code{\link{tourism_reason}} to descriptive names.
}
\examples{
data(tourism_reason_name)
tourism_reason_name
}
\references{
Tourism Research Australia: \url{https://www.tra.gov.au}
}
\seealso{
\code{\link{tourism_reason}}
}
\keyword{datasets}
