% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weather records near sightings}
\alias{Weather records near sightings}
\alias{weather}
\title{weather}
\format{
A 15274 x 9 numeric array
}
\usage{
weather
}
\description{
Daily records of precipitation, min and max temperature, covering locations and times of sightings, as available using the \code{rnoaa} package.
}
\details{
\itemize{
\item stnid, station id
\item lon, lat, elev spatial coordinates of the station
\item date, day of records
\item prcp precipitation in mm
\item tmax, tmin temperature in Celsius
}
}
\examples{

head(weather)
}
\keyword{datasets}
