% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuttlefish}
\alias{cuttlefish}
\title{cuttlefish}
\format{
A 750 x 9 numeric array
}
\usage{
cuttlefish
}
\description{
Sightings of cuttlefish downloaded from https://www.ala.org.au
}
\details{
\itemize{
\item id, record id given by Atlas
\item name, scientific name
\item lon, lat spatial coordinates of the sighting
\item date, year, month, wday, time of sighting
\item source, original provider of the record
}
}
\examples{

head(cuttlefish)
}
\keyword{datasets}
