% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_koala_data.R
\name{get_koala_data}
\alias{get_koala_data}
\title{Fetch Koala occurrence data from the Atlas of Living Australia (ALA)}
\usage{
get_koala_data(user_email)
}
\arguments{
\item{user_email}{A string containing the user's email for configuring ALA data access}
}
\value{
A tibble (data frame) containing Koala occurrence records with the following columns:
\itemize{
  \item eventDate: Date of observation (Date object)
  \item decimalLatitude: Latitude in decimal degrees (numeric)
  \item decimalLongitude: Longitude in decimal degrees (numeric)
  \item individualCount: Number of individuals observed (integer)
  \item stateProvince: Australian state where observation occurred (character)
  \item habitat_type: Type of habitat (character, from ALA's cl1048 field)
  \item status: Conservation status in the state (character)
  \item source: Legislative source for conservation status (character)
  \item assessment_year: Year of conservation status assessment (numeric)
}
}
\description{
This function retrieves Koala (Phascolarctos cinereus) occurrence records from the ALA
for the years 2023-2024, filtered by state and with quality controls. It joins the
occurrence data with conservation status information for each state.
}
\examples{
koala_data <- get_koala_data("harshithadiddige24@gmail.com")
head(koala_data)

}
