% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{stations}
\format{
A 750 x 3 numeric array
}
\usage{
stations
}
\description{
This is the nearest stations for each records in cuttlefish data
}
\details{
\itemize{
\item stnid, record id given by Atlas
\item lon, lat spatial coordinates of the sighting
}
}
\examples{

head(stations)
}
\keyword{datasets}
