% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Weather Data from Sightings Locations}
\format{
A data frame with 15,274 rows and 9 variables
}
\usage{
weather
}
\description{
Daily records of precipitation, and minimum and maximum temperature,
covering locations and times of sightings, as available using the \code{rnoaa} package.
}
\details{
\itemize{
\item stnid: Station ID
\item lon, lat, elev: Spatial coordinates and elevation of the station
\item date: Date of the record
\item prcp: Precipitation in mm
\item tmax, tmin: Temperature in Celsius
}
}
\examples{
data(weather)
head(weather)
}
\keyword{datasets}
